+++ unittests/scans/rusty_hog/gottingenhog_no_vuln copy.json
@@ -1,0 +1,1 @@
+[]
